split workspace.go by concern
