+++ internal/workspace/normalize.go
@@ -1,0 +1,10 @@
+package workspace
+
+import "path/filepath"
+
+func NormalizeRepoPath(path string) (string, error) {
+	if path == "" {
+		path = "."
+	}
+	return filepath.Abs(path)
+}
